+++ src/aui-scheduler/HISTORY.md
@@ -7,6 +7,7 @@
 * [AUI-1370](https://issues.liferay.com/browse/AUI-1370) Add ability to set an event's border color, style, and width
 * [AUI-1309](https://issues.liferay.com/browse/AUI-1309) Make aui-scheduler responsive
 * [AUI-1530](https://issues.liferay.com/browse/AUI-1530) Show red line at current time on Scheduler
+* [AUI-1531](https://issues.liferay.com/browse/AUI-1531) Scroll to current time when opening scheduler
 
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)